pyproject.toml
[project]
name = "dlt-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.7"
dependencies = [
    "faker>=37.1.0",
    "psycopg2-binary>=2.9.10",
]
